services/crawler-service/requirements.txt
feedparser
pymongo
pydantic
beautifulsoup4
requests
trafilatura
redis
python-dateutil
fastapi
uvicorn
confluent-kafka
